requirements.txt
<<<<<<< HEAD
annotated-types==0.7.0
anyio==4.6.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
distro==1.9.0
dnspython==2.7.0
Flask==2.2.5
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
MarkupSafe==2.1.2
openai==1.51.2
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
python-dotenv==0.19.1
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
=======
blinker==1.8.2
click==8.1.7
dnspython==2.7.0
Flask==2.2.5
Flask-WTF==1.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.2
>>>>>>> b8fd680d50c377cc2947af9199933fff53051b7f
Werkzeug==3.0.4
WTForms==3.0.1
